requirements.txt
faiss-cpu
sentence-transformers
faiss-cpu
markdown
python-dotenv
streamlit
pickle
datetime
json
os